Extract Base64Extension
