+++ homeassistant/helpers/template/extensions/__init__.py
@@ -1,5 +1,6 @@
 """Home Assistant template extensions."""
 
+from .base64 import Base64Extension
 from .crypto import CryptoExtension
 
-__all__ = ["CryptoExtension"]
+__all__ = ["Base64Extension", "CryptoExtension"]
