+++ tests/helpers/template/extensions/test_crypto.py
@@ -1,8 +1,6 @@
-"""Test crypto and encoding functions for Home Assistant templates."""
+"""Test cryptographic hash functions for Home Assistant templates."""
 
 from __future__ import annotations
-
-import pytest
 
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers import template
@@ -57,39 +55,4 @@ def test_sha512(hass: HomeAssistant) -> None:
     assert (
         template.Template("{{ 'Home Assistant' | sha512 }}", hass).async_render()
         == "9e3c2cdd1fbab0037378d37e1baf8a3a4bf92c54b56ad1d459deee30ccbb2acbebd7a3614552ea08992ad27dedeb7b4c5473525ba90cb73dbe8b9ec5f69295bb"
-    )
-
-
-@pytest.mark.parametrize(
-    ("value_template", "expected"),
-    [
-        ('{{ "homeassistant" | base64_encode }}', "aG9tZWFzc2lzdGFudA=="),
-        ("{{ int('0F010003', base=16) | pack('>I') | base64_encode }}", "DwEAAw=="),
-        ("{{ 'AA01000200150020' | from_hex | base64_encode }}", "qgEAAgAVACA="),
-    ],
-)
-def test_base64_encode(hass: HomeAssistant, value_template: str, expected: str) -> None:
-    """Test the base64_encode filter."""
-    assert template.Template(value_template, hass).async_render() == expected
-
-
-def test_base64_decode(hass: HomeAssistant) -> None:
-    """Test the base64_decode filter."""
-    assert (
-        template.Template(
-            '{{ "aG9tZWFzc2lzdGFudA==" | base64_decode }}', hass
-        ).async_render()
-        == "homeassistant"
-    )
-    assert (
-        template.Template(
-            '{{ "aG9tZWFzc2lzdGFudA==" | base64_decode(None) }}', hass
-        ).async_render()
-        == b"homeassistant"
-    )
-    assert (
-        template.Template(
-            '{{ "aG9tZWFzc2lzdGFudA==" | base64_decode("ascii") }}', hass
-        ).async_render()
-        == "homeassistant"
     )
